# conf/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "DIVAServices",
    "description": "Providing Document Image Analysis algorithms as WebServices",
    "termsOfService": "http://divaservices.unifr.ch/tos",
    "contact": {
      "name": "Marcel Würsch"
    }
  },
  "host": "134.21.72.48:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "list all available methods",
        "operationId": "list",
        "tags": [
          "general"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "algorithms response",
            "schema": {
              "$ref": "#/definitions/general"
            }
          }
        }
      }
    },
    "/collections": {
      "post": {
        "description": "Uploads an image to the server",
        "operationId": "addImage",
        "produces": [
          "application/json"
        ],
        "tags": [
          "collections"
        ],
        "parameters": [
          {
            "name": "images",
            "in": "body",
            "description": "Image to add on the server",
            "required": true,
            "schema": {
              "$ref": "#/definitions/images"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "upload response",
            "schema": {
              "$ref": "#/definitions/collection"
            }
          }
        }
      },
      "get": {
        "description": "list all collections",
        "produces": [
          "application/json"
        ],
        "tags": [
          "collections"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "list of collections",
            "schema": {
              "$ref": "#/definitions/collections"
            }
          }
        }
      }
    },
    "/collections/{collection}": {
      "get": {
        "description": "Get information about a collection",
        "operationId": "collectionInfo",
        "produces": [
          "application/json"
        ],
        "tags": [
          "collections"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "description": "name of the collection",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "collection information",
            "schema": {
              "$ref": "#/definitions/collectionInfo"
            }
          },
          "110": {
            "description": "collection still processing",
            "schema": {
              "$ref": "#/definitions/collectionInfo"
            }
          },
          "404": {
            "description": "collection does not exist",
            "schema": {}
          }
        }
      }
    }
  },
  "definitions": {
    "collections": {
      "type": "object",
      "required": [
        "collections"
      ],
      "properties": {
        "collections": {
          "type": "array",
          "minItems": 0,
          "uniqueItems": true,
          "items": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "url": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "images": {
      "type": "object",
      "required": [
        "images"
      ],
      "properties": {
        "images": {
          "type": "array",
          "minItems": 1,
          "uniqueItems": true,
          "items": {
            "$ref": "#/definitions/image"
          }
        }
      }
    },
    "image": {
      "type": "object",
      "required": [
        "image"
      ],
      "properties": {
        "image": {
          "type": "object",
          "required": [
            "type",
            "value"
          ],
          "properties": {
            "type": {
              "type": "string",
              "enum": [
                "url",
                "image",
                "iiif"
              ]
            },
            "value": {
              "type": "string"
            }
          }
        }
      }
    },
    "collection": {
      "type": "object",
      "required": [
        "collection"
      ],
      "properties": {
        "collection": {
          "type": "string"
        }
      }
    },
    "collectionInfo": {
      "type": "object",
      "required": [
        "statusCode",
        "statusMessage",
        "percentage"
      ],
      "properties": {
        "statusCode": {
          "type": "integer",
          "enum": [
            110,
            200
          ],
          "default": 200
        },
        "statusMessage": {
          "type": "string",
          "enum": [
            "Collection is available",
            "Downloaded x of y images"
          ],
          "default": "Collection is available"
        },
        "percentage": {
          "type": "number",
          "default": 100
        }
      }
    },
    "startExecution": {
      "type": "object",
      "required": [
        "results",
        "collection",
        "resultLink",
        "status",
        "statusCode"
      ],
      "properties": {
        "results": {
          "type": "array",
          "minItems": 1,
          "uniqueItems": true,
          "items": {
            "type": "object",
            "required": [
              "resultLink"
            ],
            "properties": {
              "resultLink": {
                "type": "string"
              }
            }
          }
        },
        "collection": {
          "type": "string"
        },
        "resultLink": {
          "type": "string",
          "format": "uri"
        },
        "status": {
          "type": "string"
        },
        "statusCode": {
          "type": "number"
        }
      }
    },
    "general": {
      "type": "array",
      "items": {
        "description": "schema for an algorithm overview entry",
        "type": "object",
        "required": [
          "name",
          "description",
          "url"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "minLength": 5,
            "maxLength": 255
          },
          "type": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "url": {
            "type": "string",
            "format": "uri"
          }
        },
        "additionalProperties": false
      }
    }
  }
}